Apply review feedback for GPUI steps, docs links, trybuild staging
Document the `test_function_name()` call in the GPUI counter example without
asserting `None`. Fix migration guide links in crate READMEs to reach the
repo-level docs path.

Share trybuild file-copy helpers via a hidden `trybuild_staging` module on
`rstest-bdd-harness` so Tokio and GPUI macro compile tests stay aligned.
