+++ crates/rstest-bdd-harness/src/lib.rs
@@ -10,6 +10,8 @@ mod runner;
 mod std_harness;
 #[cfg(test)]
 pub(crate) mod test_utils;
+#[doc(hidden)]
+pub mod trybuild_staging;
 
 pub use adapter::HarnessAdapter;
 pub use error::{HarnessError, HarnessResult};
